spotify/requirements.txt
spotipy
pyserial
python-dotenv
